SetVersion fixed, mostly
